backend/btc: move rates to backend
